requirements.txt
openai==1.54.4
python-dotenv==1.0.1
